requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
librouteros==3.4.1
MarkupSafe==3.0.2
paho-mqtt==2.1.0
PyMySQL==1.1.1
toml==0.10.2
Werkzeug==3.1.3s